tg/requirements.txt
aiogram==3.13.0
aiohttp>=3.9.0
python-dotenv==1.0.1
pytz==2024.1
